Consolidate api directory
Move operator-controller APIs into apis/operator-controller
Move catalogd APIs into apis/catalogd

There's a conflict in types, both have a CatalogSource type that we may
not be able to easily rename, so the simplest solution was to keep the
two APIs serparate
