+++ Tiltfile
@@ -6,7 +6,7 @@ operator_controller = {
     'binaries': {
         './cmd/operator-controller': 'operator-controller-controller-manager',
     },
-    'deps': ['api', 'cmd/operator-controller', 'internal', 'pkg', 'go.mod', 'go.sum'],
+    'deps': ['api/operator-controller', 'cmd/operator-controller', 'internal/operator-controller', 'internal/shared', 'go.mod', 'go.sum'],
     'starting_debug_port': 30000,
 }
 deploy_repo('operator-controller', operator_controller, '-tags containers_image_openpgp')
@@ -17,7 +17,7 @@ catalogd = {
     'binaries': {
         './catalogd/cmd/catalogd': 'catalogd-controller-manager',
     },
-    'deps': ['catalogd/api', 'catalogd/cmd/catalogd', 'catalogd/internal', 'catalogd/pkg', 'go.mod', 'go.sum'],
+    'deps': ['api/catalogd', 'catalogd/cmd/catalogd', 'internal/catalogd', 'internal/shared', 'go.mod', 'go.sum'],
     'starting_debug_port': 20000,
 }
