+++ api/catalogd/doc.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-//go:generate apiregister-gen --input-dirs ./... -h ../../boilerplate.go.txt
+//go:generate apiregister-gen --input-dirs ./... -h ../../hack/boilerplate.go.txt
 
 //
 // +domain=operatorframework.io
