+++ internal/catalogd/controllers/core/clustercatalog_controller_test.go
@@ -21,7 +21,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	catalogdv1 "github.com/operator-framework/operator-controller/catalogd/api/v1"
+	catalogdv1 "github.com/operator-framework/operator-controller/api/catalogd/v1"
 	"github.com/operator-framework/operator-controller/internal/catalogd/storage"
 	imageutil "github.com/operator-framework/operator-controller/internal/shared/util/image"
 )
