+++ internal/operator-controller/catalogmetadata/client/client_test.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/operator-framework/operator-registry/alpha/declcfg"
 
-	catalogd "github.com/operator-framework/operator-controller/catalogd/api/v1"
+	catalogd "github.com/operator-framework/operator-controller/api/catalogd/v1"
 	catalogClient "github.com/operator-framework/operator-controller/internal/operator-controller/catalogmetadata/client"
 )
