+++ internal/operator-controller/catalogmetadata/filter/successors.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/operator-framework/operator-registry/alpha/declcfg"
 
-	ocv1 "github.com/operator-framework/operator-controller/api/v1"
+	ocv1 "github.com/operator-framework/operator-controller/api/operator-controller/v1"
 	"github.com/operator-framework/operator-controller/internal/shared/util/filter"
 )
